genesis: add Bridge Vault UTXO Seed (14 outputs) + fix u128 overflow
- Split Bridge Seed Fund into:
  - Slot 13: 400M ZION (account model)
  - Slot 14: 100M ZION (UTXO coinbase, 6 outputs) for bridge vault liquidity
- Genesis block now builds a hybrid block: 13 account txs + 1 UTXO tx
- Fixed utxo_balance() overflow: changed return type u64 -> u128
- Fixed format_flowers_as_zion() to accept u128
- Fixed websocket u128 serialization panic in WsConfirmation
- Updated all docs (README, AGENTS, whitepapers, audits, guides)
- Hard-reset both local PC and Edge server with new genesis
- Edge pool rebuilt and restarted
- Added Edge operational notes to AGENTS.md

+++ V3/L1/core/src/lib.rs
@@ -1230,7 +1230,7 @@ impl NodeRuntime {
         self.chain_state.height
     }
 
-    pub fn utxo_balance(&self, address: &str) -> u64 {
+    pub fn utxo_balance(&self, address: &str) -> u128 {
         self.chain_state.utxo_balance(address)
     }
 
@@ -1891,11 +1891,11 @@ impl ChainState {
     }
 
     /// Compute balance for a `zion1...` address by summing unspent UTXO outputs.
-    fn utxo_balance(&self, address: &str) -> u64 {
+    fn utxo_balance(&self, address: &str) -> u128 {
         self.utxo_set()
             .values()
             .filter(|u| u.address == address)
-            .map(|u| u.amount)
+            .map(|u| u.amount as u128)
             .sum()
     }
